Checking in some Java 5 changes
git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@4983 51baf565-9d33-0410-a72c-fc3788e3496d

+++ pmd/src/net/sourceforge/pmd/Report.java
@@ -160,7 +160,7 @@ public class Report {
             if (!summary.containsKey(name)) {
                 summary.put(name, NumericConstants.ZERO);
             }
-            Integer count = (Integer) summary.get(name);
+            Integer count = summary.get(name);
             summary.put(name, new Integer(count.intValue() + 1));
         }
         return summary;
